Handle warnings, make the buttons work
The hookButtons function now checks for the current page before trying to hook the addItem button and the info button, so as to avoid the error spam from trying to add an event listener to a button that doesn't exist. This means that the hookButtons function can now also be used to hook the buttons in the Today and Upcoming pages, so that we retain their functionality.