Make the project dropdown default to the current project
The project dropdown in the addItem modal was always defaulting to none. That worked, but was annoying and could easily lead to someone posting a task on the wrong project. This was changed and now the addItemModal checks the current page and defaults that option to whatever page we are on.