Fixed bug of the duplicate event listener on edit
Behavior: When on the Today or Upcoming pages the edit button was given two click event listeners, causing two modals to pop up.
Cause: On the swapPage function the hookButtons() is called by default, but was also explicitly called every time the user clicked the today and upcoming buttons. eg.

      case "Today" :
        ...
        hookButtons();
        break;
      case "Upcoming" :
        ...
        hookButtons();
        break;
  ...
  hookButtons();

Removing the hookButtons() from those cases fixes the issue.