Fix bug that causes content to disappear when switching themes
When switching themes the e.target could refer to either the img or the parent button, depending where the user clicked. This caused the content to disappear if the user clicked at the corner of that button. This is fixed by creating a direct global reference to themeButton and using that to add and remove the svgs from.